requirements.txt
newrelic==2.60.0.46
netaddr==0.7.13
Cython==0.20.2
Django==1.6.5
Hoover==0.6.0
#MySQL-python==1.2.5
amqp==1.0.13
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.34
boto==2.32.1
celery==3.0.25
celery-with-mongodb==3.0
certifi==14.05.14
dj-database-url==0.3.0
django-bootstrap3==4.11.0
django-celery==3.0.21
docutils==0.12
gunicorn==19.1.1
kombu==2.5.16
logilab-astng==0.24.3
logilab-common==0.62.1
lxml==3.3.6
numpy==1.9.0
numexpr==2.4
pandas==0.15.0
scipy==0.14.0
scikit-learn==0.15.2
pyinterval==1.0.0
pip==1.5.6
psycopg2==2.5.4
pycrypto==2.6.1
pymongo==2.8
python-dateutil==2.2
python-memcached==1.57
pylibmc==1.5.0
pytz==2014.7
requests==2.4.0
-e git+https://github.com/gnanarepo/sfdc-python-toolkit.git#egg=salesforce-python-toolkit
six==1.7.3
suds==0.3.9
wsgiref==0.1.2
salesforce-bulk==1.0.7
openpyxl==2.2.2
south==1.0.2
-e git+https://github.com/gnanarepo/py-schema.git#egg=py-schema
